+++ packages/e2e/test/integration/first.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: avoid_relative_lib_imports
 
 import 'package:e2e/annotation.dart';
+import 'package:other/nested/file.dart' as other_file;
 
 import '../../lib/e2e/../e2e/b.dart' as not_normalized_path_prefix;
 import './a.dart' as prefix;
@@ -23,3 +24,8 @@ void prefixToRelativePath([
   Object value3 = const direct_path_prefix.C(),
   Object value6 = const relative_reexport.F(),
 ]) {}
+
+@E2E()
+void assetUri({
+  Object value = const other_file.File(),
+}) {}