+++ packages/e2e/test/integration/first.e2e.dart
@@ -12,6 +12,7 @@ import 'a.dart' as _i2;
 import '../../lib/e2e/b.dart' as _i3;
 import 'c.dart' as _i4;
 import 'f.dart' as _i5;
+import 'package:other/nested/file.dart' as _i6;
 
 void simpleFromTestE2e([
   Object value = const _i1.Foo(),
@@ -22,5 +23,9 @@ void prefixToRelativePathE2e([
   Object value2 = const _i3.B(),
   Object value3 = const _i4.C(),
   Object value6 = const _i5.F(),
+]) =>
+    throw UnimplementedError();
+void assetUriE2e([
+  Object value = const _i6.File(),
 ]) =>
     throw UnimplementedError();