+++ packages/other/lib/nested/file.dart
@@ -1,0 +1,6 @@
+/// Check for "package:other/nested/file.dart" import, as they arrive as
+/// "asset:other/lib/nested/file.dart" in the generated code.
+class File {
+  /// File
+  const File();
+}